pyproject.toml
[tool.poetry]
name = "tableau-view-extractor"
version = "0.1.0"
description = "Application that provides an interface for extracting data from a Tableau Server instance."
authors = ["Todd Birchard <todd@hackersandslackers.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "tableau_view_extractor"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.2"
Flask = "^2.2.3"
flask-redis = "^0.4.0"
Flask-SQLAlchemy = "^3.0.3"
gunicorn="*"
pymysql="*"
pandas = "^1.5.3"
jsmin = "^3.0.1"
libsass = "^0.22.0"
tableauserverclient = "^0.25"
black = "*"
isort = "*"
flake8 = "*"
pylint = "*"
uwsgi = "^2.0.21"
python-dotenv = "^1.0.0"


[tool.poetry.scripts]
run = "wsgi:app"

[tool.poetry.urls]
issues = "https://github.com/toddbirchard/tableau-view-extractor/issues"

[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry>=1.3.1"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
src_paths = ["tableau_view_extractor", "config"]

[tool.black]
line-length = 80

[tool.pylint.'MESSAGES CONTROL']
disable="C0103,C0301,W0703,W0621"